backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
django-filter==23.5
drf-spectacular==0.27.1
python-decouple==3.8
Pillow==10.2.0
